refactor(post-comment-thread): rename compnents fix minior display error
