--- apps/ourvoice-app/src/graphql/mutations/votePost.ts
@@ -1,19 +1,0 @@
-import gql from 'graphql-tag'
-
-export const VOTE_POST_MUTATION = gql`
-  mutation VotePost($data: VoteCreateInput!) {
-    createVote(data: $data) {
-      id
-      voteType
-      user {
-        id
-      }
-      post {
-        id
-      }
-      comment {
-        id
-      }
-    }
-  }
-`
